Improve CodeStyle and CodeQuality
There are several areas which were pointed out by coursemates to be improved for a better code readability and code quality.

Let's change the variable name to adhere to camelcase naming and change the structure of method call echo() by using switch statement method and initialized new Task object in the smallest scope.